Use any installed VC compiler to build the Python package.
The distutils C compiler support is intended for building Python
extension modules. These modules need to be built with the same
compiler version than the Python interpreter was built in Windows. For
Python 2.7 this would be Visual C++ from 2007.

Distorm3 builds a native DLL which is used by ctypes in Python so it is
not bound by these restrictions.

Change the setup to use any installed Visual C++ and build the DLL using
the solution file.

This reduces configuration duplication and removes the need to find a
10-year-old compiler.